+++ lib/components/PlayerControls.dart
@@ -18,19 +18,15 @@ class PlayerControls extends StatelessWidget {
   const PlayerControls({
     super.key,
     required this.track,
-    // required this.trackInfo,
     required this.playSeekBackward,
     required this.playSeekForward,
-    // required this.setTrack,
     required this.togglePause,
     required this.isPlaying,
     required this.isPaused,
   });
   final Track track;
-  // final TrackInfo? trackInfo;
   final VoidCallback playSeekBackward;
   final VoidCallback playSeekForward;
-  // final void Function(Track? track, {bool play, bool notify}) setTrack;
   final VoidCallback togglePause;
   final bool isPlaying;
   final bool isPaused;
@@ -41,24 +37,14 @@ class PlayerControls extends StatelessWidget {
       spacing: 0,
       mainAxisAlignment: MainAxisAlignment.center,
       children: [
-        PlayerBackwardControl(
-            track: track,
-            // trackInfo: trackInfo,
-            playSeekBackward: this.playSeekBackward),
+        PlayerBackwardControl(track: track, playSeekBackward: this.playSeekBackward),
         PlayerPlayControl(
-            track: track,
-            // trackInfo: trackInfo,
-            togglePause: this.togglePause,
-            isPlaying: this.isPlaying,
-            isPaused: this.isPaused),
-        PlayerForwardControl(
-            track: track,
-            // trackInfo: trackInfo,
-            playSeekForward: this.playSeekForward),
+            track: track, togglePause: this.togglePause, isPlaying: this.isPlaying, isPaused: this.isPaused),
+        PlayerForwardControl(track: track, playSeekForward: this.playSeekForward),
         PlayerFavoriteControl(
           track: track,
-          // trackInfo: trackInfo,
         ),
+        PlayerHideControl(),
       ],
     );
   }
@@ -68,27 +54,22 @@ class PlayerPlayControl extends StatelessWidget {
   const PlayerPlayControl({
     super.key,
     required this.track,
-    // required this.trackInfo,
     required this.togglePause,
     required this.isPlaying,
     required this.isPaused,
   });
   final Track track;
-  // final TrackInfo? trackInfo;
   final VoidCallback togglePause;
   final bool isPlaying;
   final bool isPaused;
 
   @override
   Widget build(BuildContext context) {
-    // final appState = context.watch<AppState>();
     final theme = Theme.of(context);
     final AppColors appColors = theme.extension<AppColors>()!;
-    // final colorScheme = theme.colorScheme;
 
     final mainColor = appColors.brandColor;
 
-    // final isPlaying = appState.isPlaying && !appState.isPaused;
     final showPlaying = this.isPlaying && !this.isPaused;
 
     return Stack(
@@ -125,19 +106,15 @@ class PlayerBackwardControl extends StatelessWidget {
   const PlayerBackwardControl({
     super.key,
     required this.track,
-    // required this.trackInfo,
     required this.playSeekBackward,
   });
   final Track track;
-  // final TrackInfo? trackInfo;
   final VoidCallback playSeekBackward;
 
   @override
   Widget build(BuildContext context) {
-    // final appState = context.watch<AppState>();
     final theme = Theme.of(context);
     final AppColors appColors = theme.extension<AppColors>()!;
-    // final colorScheme = theme.colorScheme;
 
     return IconButton(
       icon: Icon(
@@ -155,19 +132,15 @@ class PlayerForwardControl extends StatelessWidget {
   const PlayerForwardControl({
     super.key,
     required this.track,
-    // required this.trackInfo,
     required this.playSeekForward,
   });
   final Track track;
-  // final TrackInfo? trackInfo;
   final VoidCallback playSeekForward;
 
   @override
   Widget build(BuildContext context) {
-    // final appState = context.watch<AppState>();
     final theme = Theme.of(context);
     final AppColors appColors = theme.extension<AppColors>()!;
-    // final colorScheme = theme.colorScheme;
 
     return IconButton(
       icon: Icon(
@@ -185,19 +158,16 @@ class PlayerFavoriteControl extends StatelessWidget {
   const PlayerFavoriteControl({
     super.key,
     required this.track,
-    // required this.trackInfo,
   });
   final Track track;
-  // final TrackInfo? trackInfo;
 
   @override
   Widget build(BuildContext context) {
     final appState = context.watch<AppState>();
     final theme = Theme.of(context);
     final AppColors appColors = theme.extension<AppColors>()!;
-    // final colorScheme = theme.colorScheme;
 
-    final isFavorite = appState.isFavoriteTrackId(track.id); // trackInfo?.favorite ?? false;
+    final isFavorite = appState.isFavoriteTrackId(track.id);
 
     return IconButton(
       icon: Icon(
@@ -209,6 +179,31 @@ class PlayerFavoriteControl extends StatelessWidget {
       onPressed: () {
         final setFavorite = !isFavorite;
         appState.setFavorite(track.id, setFavorite);
+      },
+    );
+  }
+}
+
+class PlayerHideControl extends StatelessWidget {
+  const PlayerHideControl({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    final appState = context.watch<AppState>();
+    final theme = Theme.of(context);
+    final AppColors appColors = theme.extension<AppColors>()!;
+
+    return IconButton(
+      icon: Icon(
+        Icons.close,
+        color: appColors.brandColor,
+      ),
+      alignment: Alignment.center,
+      padding: EdgeInsets.all(0.0),
+      onPressed: () {
+        appState.hidePlayer();
       },
     );
   }